Add in jumping and ducking
This was based off a patch that Zack gave to me.

You will want to tweak the constant jumpTimeLength, and also the functions used to map the factor in the jump code (line 150)
Right now, we map the first half of the animation to the rise, and the second half to the fall. a different mapping might be better.